# v2/docs/swagger.yaml
basePath: /api/v2
definitions:
  github_com_sabbatD_srest-api_internal_lib_todoConfig.Meta:
    properties:
      totalAmount:
        type: integer
    type: object
  github_com_sabbatD_srest-api_internal_lib_todoConfig.MetaResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.Todo'
        type: array
      info:
        $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.TodoInfo'
      meta:
        $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.Meta'
    type: object
  github_com_sabbatD_srest-api_internal_lib_todoConfig.Todo:
    properties:
      created:
        type: string
      description:
        type: string
      executor:
        type: string
      id:
        type: integer
      isDone:
        type: boolean
      title:
        type: string
    type: object
  github_com_sabbatD_srest-api_internal_lib_todoConfig.TodoInfo:
    properties:
      all:
        type: integer
      completed:
        type: integer
      inWork:
        type: integer
    type: object
  github_com_sabbatD_srest-api_internal_lib_todoConfig.TodoRequest:
    properties:
      description:
        type: string
      executor:
        type: string
      isDone:
        type: boolean
      title:
        type: string
    type: object
...
host: easydev.club
info:
  contact:
    email: s4bb4t@yandex.ru
    name: s4bb4t
  description: This is a RESTful API service for EasyDev. It provides various user
    management functionalities such as user registration, authentication, profile
    updates, and admin operations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: sAPI
  version: v0.3.2
paths:
...
  /todos:
    get:
      description: Retrieves all tasks with optional filtering by status (e.g., completed
        or in-progress).
      parameters:
        - description: 'Filter tasks by status: all, completed, or inWork'
          in: query
          name: filter
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Tasks retrieved successfully.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.MetaResponse'
        "500":
          description: Internal server error.
          schema:
            type: string
      summary: Retrieve all tasks
      tags:
        - todo
    post:
      consumes:
        - application/json
      description: Creates a new task by accepting a JSON payload with the task's
        details.
      parameters:
        - description: Task data for creating a new task
          in: body
          name: UserData
          required: true
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.TodoRequest'
      produces:
        - application/json
      responses:
        "200":
          description: Task successfully created, returns the created task.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.Todo'
        "400":
          description: Invalid request body or missing/incorrect fields.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      summary: Create a new task
      tags:
        - todo
  /todos/{id}:
    delete:
      description: Deletes a task by its ID from the URL.
      parameters:
        - description: ID of the task to delete
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Task deleted successfully.
          schema:
            type: string
        "400":
          description: Invalid or missing task ID.
          schema:
            type: string
        "404":
          description: Task not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      summary: Delete a task by ID
      tags:
        - todo
    get:
      description: Retrieves a specific task by its ID from the URL.
      parameters:
        - description: ID of the task to retrieve
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Task retrieved successfully.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.Todo'
        "400":
          description: Invalid or missing task ID.
          schema:
            type: string
        "404":
          description: Task not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      summary: Retrieve a task by ID
      tags:
        - todo
    put:
      consumes:
        - application/json
      description: Updates an existing task by accepting a JSON payload with the updated
        task details.
      parameters:
        - description: ID of the task to update
          in: path
          name: id
          required: true
          type: integer
        - description: Updated task data
          in: body
          name: UserData
          required: true
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.TodoRequest'
      produces:
        - application/json
      responses:
        "200":
          description: Task updated successfully, returns the updated task.
          schema:
            $ref: '#/definitions/github_com_sabbatD_srest-api_internal_lib_todoConfig.Todo'
        "400":
          description: Invalid request body, missing/incorrect fields, or invalid
            ID.
          schema:
            type: string
        "404":
          description: Task not found.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      summary: Update an existing task
      tags:
        - todo
...
schemes:
  - https
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token required for accessing protected routes. Format:
      "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"